chore: fixing bugs to resolve intake and ingest builds
