+++ docker/Dockerfile.ingest
@@ -7,7 +7,7 @@ RUN yum -y upgrade
 RUN yum -y install gcc ca-certificates
 
 # Install python deps deps
-RUN pip3 install stac_fastapi.opensearch stac_fastapi.types
+RUN pip3 install stac_fastapi.opensearch stac_fastapi.types python-json-logger
 
 # Copy the function code to the LAMBDA_TASK_ROOT directory
 ADD . ${LAMBDA_TASK_ROOT}
